requirements.txt
# Django Framework
Django==5.0.2
djangorestframework==3.14.0

# Database
psycopg2-binary==2.9.9

# Environment variables
python-decouple==3.8

# Image handling
Pillow==10.2.0

# Email
django-anymail==11.0

# Rich text editor
django-ckeditor==6.7.0

# Forms
django-crispy-forms>=2.3
crispy-bootstrap5==2024.10

# Security
django-cors-headers==4.3.1

# Date and time
python-dateutil==2.8.2

# Production server
gunicorn==21.2.0
whitenoise==6.6.0

# Development tools
django-debug-toolbar==4.3.0
